Assign MediaEvents MessageType of Media
Previously MediaEvents were created with a MessageType of Unknown. This MessageType was not an issue for the core SDK, since we do not upload these events to the MParticle servers, but it did introduce a bug when forwarding these events to Kits, which resulted in all MediaEvents getting filtered before they could be forwarded to the individual kits. In the filter stage, we remove events with MessageType == Unknown as malformed, so updating these events to their proper MessageType will resolve the issue